perf(collation): build sort keys in one pass with an ASCII fast path
Profiling the final serialization pass on 4650 messages across four
locales, per catalog file of 4950 entries:

  parse_po      ~1.2 ms
  metadata      ~0.0 ms
  sort          ~2.5 ms
  stringify     ~0.5 ms

So the sort was the largest single piece, at roughly 500 ns per entry
for two keys — far more than a table lookup and two pushes should cost.

Two things turned out not to be the problem. Cloning the raw message
into the sort key measures the same either way, so it stays for a
deterministic tie-break. And packing the three levels into one buffer by
walking the string once per level made it *worse*, 2.5 ms to 3.4 ms:
per-character work dominates, not allocation count, and three passes
means three UTF-8 decodes and three table lookups.

What actually helps is doing less per character. The levels are now
filled in a single pass, and ASCII — which is nearly all of a source
catalog — skips UTF-8 decoding, the combining-mark check, the expansion
branch and the out-of-range table fallback entirely. Entries without a
gettext context, the common case, no longer allocate a key at all.

  sort        2.5 ms -> 1.86 ms   (-26%)
  extract     92 ms  -> 88 ms

The key is a single byte buffer with level boundaries rather than three
vectors, so comparisons run as memcmp; primary weights are ranks in a
139-character repertoire and fit in a byte, which the generator now
enforces.

Verified unchanged: 4922 msgids extracted through the release binary,
zero at a different position than `Intl.Collator("en-US")` puts them.

What remains is structural. parse_po plus stringify_po is about half the
final pass, and it exists only because ferrocat renders in code-point
order and its render options cannot express a collation. Removing that
round trip needs an upstream change, not a local one.

+++ crates/palamedes/src/collation_table.rs
@@ -6,7 +6,7 @@
 /// Everything the key builder needs for one code point.
 pub(crate) struct Row {
     /// Rank of the base letter in CLDR root order, or 0 when uncovered.
-    pub(crate) primary: u16,
+    pub(crate) primary: u8,
     /// Combining mark contributed to the secondary level, or 0 for none.
     pub(crate) secondary: u32,
     /// Whether the character is uppercase, which is a tertiary difference.
